requirements.txt
click
vpype>=1.9,<2.0
numpy
freetype-py